requirements.txt
beautifulsoup4==4.11.1
black==22.6.0
bs4==0.0.1
click==8.1.3
colorama==0.4.5
mypy-extensions==0.4.3
mysql-connector-python==8.0.30
pathspec==0.10.1
platformdirs==2.5.2
protobuf==3.20.1
psycopg2-binary==2.9.3
python-dotenv==0.21.0
soupsieve==2.3.2.post1
tomli==2.0.1
typing_extensions==4.3.0
